route new flags through featureflags interface
